scons: build: override open file resource limit to avoid SCons performance penalty
Reduce maximum number of open files resource limit.
Internally, SCons contains a loop which invokes close(2)
for 0..N files where N is the maximum number of open files.
The loop is called every time a new process is created (e.g. from a builder).
This loop becomes a serious performance issue when the limit
is set to a high value, e.g. 1 million on an Ubuntu 18.04 Jenkins slave.
Therefore, we override the limit to prevent the potential performance penalty.