arm: raspberry: add support for newer Qemu versions using QEMU_BUILD build variable
Unfortunately, recent Qemu versions do not support the original Raspberry Pi
machine anymore (-M raspi). Only raspberry pi 2, 3 and newer are supported.
In order to partially emulate and test code for the raspberry pi 1 under Qemu,
the following build variable QEMU_BUILD ensures that the code will be compiled for the
ARM1177 processor with ARMv6 architecture, but use all BCM2836 specifics such
as the I/O base address of the raspberry pi 2 for Qemu. Therefore, when setting
QEMU_BUILD to True the compiled image will not work on a real Raspberry Pi 1.