kernel/API/VMCtl: request allocation of absolute range of memory in ReserveMem
In the current implementation, VMCtl(ReserveMem) would only search
for the first available block of memory of sufficient size. However,
for the secondary cores, the range of memory that needs to be reserved is
fixed to a pre-calculated range of physical memory. Without this guarantee,
it is possible that the boot core and secondary cores have overlapping
free range(s) of memory pages.

This commit ensures that VMCtl() always requests allocation of
the given range of physical memory pages.