lib/libipc/ChannelClient: perform explicit ProcessCtl(Schedule) when VMShare returns TemporaryUnavailable
In some cases when ChannelClient connects to another process with VMShare, it is possible
that the other process is still removing a previous VMShare-entry when a PID
is re-used. In such case, VMShare returns API::Result with value TemporaryUnavailable.
After ChannelClient has retried 16 times to connect, it gives up and returns an error.

However if the other process does not get sufficient CPU time to actually complete
the VMShare-entry cleanup, this error can be triggered as well. Resolve the problem
by invoking ProcessCtl(Schedule) explicitly after VMShare returns TemporaryUnavailable.