server/network/sun8i/Sun8iEmac: write frame descriptor base address before activating DMA
When activating DMA before the frame descriptor base address is written, there is a small
chance depending on timing that the EMAC's DMA engine reads the frame descriptor base address
before it is fully written to RAM by the processor. This can lead to the
"receive descriptor unavailable" interrupt flag to be set and the DMA engine will be halted.